+++ codigo/PQ.h
@@ -35,6 +35,6 @@ bool PQ_is_empty(PQ *pq);
  */
 int PQ_size(PQ *pq);
 
-void PQ_decrementaChave(PQ *pq, tVertice *v, int novaDist);
+void PQ_decrementaChave(PQ *pq, tVertice *v, float novaDist);
 
 #endif
